Added interface IParentNode; implemented a parser in the RootNode class
IParentNode implements INode, adds a list and a method for adding to that list. The 
list itself contains INodes. MainDomTreee no longer stores list of all the rootlevel
nodes. For that purpose was created a new class of nodes called RootNode. The
MainDomTree contains only a refence to single instance of the RootNode class.
That class can build a tree by parsing a document, passed as a string