Fixed a bug in PageInfo. Added AspNetDom Foldings.
Fixed an issue in PageInfo which didn't let the foreach statement to
iterate through the tree. Added foldings for the nodes from
AspNet.StateEngine.AspNetDom.cs and Xml.StateEngine.XDom. The Document
outline is now built on the AST from the StateEngine parser. The
RootNode propery in AspNetParsedDocument is still to satisfy
dependencies in the addin.